IGNITE-16447: added LifecycleAware support for local listener.
